LangChain text/requirements.txt
flask>=3.0.0
langchain>=0.3.4
langchain-community>=0.3.3
google-genai>=1.0.0
python-dotenv>=1.0.1
flask_cors